Make tests faster again.
