+++ phpmainthread_test.go
@@ -12,7 +12,6 @@ import (
 	"testing"
 	"time"
 
-	"github.com/dunglas/frankenphp/internal/phpheaders"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -150,21 +149,6 @@ func TestTransitionThreadsWhileDoingRequests(t *testing.T) {
 	Shutdown()
 }
 
-// Note: this test is here since it would break compilation when put into the phpheaders package
-func TestAllCommonHeadersAreCorrect(t *testing.T) {
-	fakeRequest := httptest.NewRequest("GET", "http://localhost", nil)
-
-	for header, phpHeader := range phpheaders.CommonRequestHeaders {
-		// verify that common and uncommon headers return the same result
-		expectedPHPHeader := phpheaders.GetUnCommonHeader(header)
-		assert.Equal(t, phpHeader+"\x00", expectedPHPHeader, "header is not well formed: "+phpHeader)
-
-		// net/http will capitalize lowercase headers, verify that headers are capitalized
-		fakeRequest.Header.Add(header, "foo")
-		_, ok := fakeRequest.Header[header]
-		assert.True(t, ok, "header is not correctly capitalized: "+header)
-	}
-}
 func TestFinishBootingAWorkerScript(t *testing.T) {
 	workers = nil
 	logger = slog.New(slog.NewTextHandler(io.Discard, nil))
